Refactor LLM model into smaller modules for better code organization
- Split large candle_model.rs (1685 lines) into smaller modules (<400 lines)
- Extract KVCache, TransformerLayer into separate files
- Extract forward pass and generation logic into dedicated modules
- Remove debug statements from all files
- Simplify comments to follow "why not what" guideline
- Maintain code cleanliness and structure per CLAUDE.md requirements
