+++ crates/arkavo-llm/src/candle/mod.rs
@@ -1,0 +1,7 @@
+pub mod forward;
+pub mod generation;
+pub mod kv_cache;
+pub mod transformer_layer;
+
+// Re-export TransformerLayer for use in the main module
+pub use transformer_layer::TransformerLayer;
